requirements.txt
django
django-compressor
django-extensions
django-grappelli
django-libsass
django-model-utils
django-redis
psycopg2-binary
redis

django-activecollab-digger
django-apscheduler
django-kdl-ldap
django-polymorphic
django-rest-framework
jsonfield
pymongo
requests_oauthlib
textblob
